refactor: rename uils → lib
