+++ packages/takudev-web/lib/getCustomColorPropertySettingText.ts
